make event_custom_ops opaque to avoid leaking lua dependency
