+++ matron/src/event_types.h
@@ -3,7 +3,6 @@
 #include "oracle.h"
 #include "osc.h"
 #include <stdint.h>
-#include <lua.h>
 
 typedef enum {
     // unused (do not remove)
@@ -316,14 +315,8 @@ struct event_softcut_position {
     float pos;
 };
 
-typedef void (*event_custom_weave_op_t)(lua_State *lvm, void *value, void *context);
-typedef void (*event_custom_free_op_t)(void *value, void *context);
-
-struct event_custom_ops {
-    const char *type_name;
-    event_custom_weave_op_t weave;
-    event_custom_free_op_t free;
-};
+// forward declaration to hide scripting layer dependencies
+struct event_custom_ops;
 
 struct event_custom {
     struct event_common common;
