+++ matron/src/events.c
@@ -17,6 +17,7 @@
 #include "weaver.h"
 
 #include "event_types.h"
+#include "event_custom.h"
 
 //----------------------------
 //--- types and variables
@@ -309,7 +310,7 @@ static void handle_event(union event_data *ev) {
         w_handle_softcut_position(ev->softcut_position.idx, ev->softcut_position.pos);
         break;
     case EVENT_CUSTOM:
-        w_handle_custom_weave(ev->custom.ops->weave, ev->custom.value, ev->custom.context);
+        w_handle_custom_weave(&(ev->custom));
         break;
     } /* switch */
