+++ matron/src/weaver.c
@@ -33,6 +33,7 @@
 #include "device_midi.h"
 #include "device_monome.h"
 #include "events.h"
+#include "event_custom.h"
 #include "hello.h"
 #include "i2c.h"
 #include "lua_eval.h"
@@ -2094,10 +2095,10 @@ void w_handle_system_cmd(char *capture) {
     l_report(lvm, l_docall(lvm, 1, 0));
 }
 
-void w_handle_custom_weave(event_custom_weave_op_t op, void *value, void *context) {
+void w_handle_custom_weave(struct event_custom *ev) {
     // call the externally defined `op` function passing in the current lua
     // state
-    op(lvm, value, context);
+    ev->ops->weave(lvm, ev->value, ev->context);
 }
 
 // helper: set poll given by lua to given state
